Renamed nested 'permissions' endpoints to 'permission-assignments'
